Add Relay/Connection specific MetadataGenerator and RecordMerger
