+++ apollo-normalized-cache-api-incubating/src/commonMain/kotlin/com/apollographql/apollo3/cache/normalized/api/RecordMerger.kt
@@ -94,3 +94,59 @@ class FieldRecordMerger(private val fieldMerger: FieldMerger) : RecordMerger {
     ) to changedKeys
   }
 }
+
+@ApolloExperimental
+val ConnectionRecordMerger = FieldRecordMerger(ConnectionFieldMerger)
+
+private object ConnectionFieldMerger : FieldRecordMerger.FieldMerger {
+  @Suppress("UNCHECKED_CAST")
+  override fun mergeFields(existing: FieldRecordMerger.FieldInfo, incoming: FieldRecordMerger.FieldInfo): FieldRecordMerger.FieldInfo {
+    val existingStartCursor = existing.metadata["startCursor"] as? String
+    val existingEndCursor = existing.metadata["endCursor"] as? String
+    val incomingStartCursor = incoming.metadata["startCursor"] as? String
+    val incomingEndCursor = incoming.metadata["endCursor"] as? String
+    val incomingBeforeArgument = incoming.metadata["before"] as? String
+    val incomingAfterArgument = incoming.metadata["after"] as? String
+
+    return if (incomingBeforeArgument == null && incomingAfterArgument == null) {
+      // Not a pagination query
+      incoming
+    } else if (existingStartCursor == null || existingEndCursor == null) {
+      // Existing is empty
+      incoming
+    } else if (incomingStartCursor == null || incomingEndCursor == null) {
+      // Incoming is empty
+      existing
+    } else {
+      val existingValue = existing.value as Map<String, Any?>
+      val existingList = existingValue["edges"] as List<*>
+      val incomingList = (incoming.value as Map<String, Any?>)["edges"] as List<*>
+
+      val mergedList: List<*>
+      val newStartCursor: String
+      val newEndCursor: String
+      if (incomingAfterArgument == existingEndCursor) {
+        mergedList = existingList + incomingList
+        newStartCursor = existingStartCursor
+        newEndCursor = incomingEndCursor
+      } else if (incomingBeforeArgument == existingStartCursor) {
+        mergedList = incomingList + existingList
+        newStartCursor = incomingStartCursor
+        newEndCursor = existingEndCursor
+      } else {
+        // We received a list which is neither the previous nor the next page.
+        // Handle this case by resetting the cache with this page
+        mergedList = incomingList
+        newStartCursor = incomingStartCursor
+        newEndCursor = incomingEndCursor
+      }
+
+      val mergedFieldValue = existingValue.toMutableMap()
+      mergedFieldValue["edges"] = mergedList
+      FieldRecordMerger.FieldInfo(
+          value = mergedFieldValue,
+          metadata = mapOf("startCursor" to newStartCursor, "endCursor" to newEndCursor)
+      )
+    }
+  }
+}
